requirements.txt
pandas
numpy
matplotlib
seaborn
streamlit
google-api-python-client
python-dotenv
scikit-learn
